add flow files

+++ streampark-console/streampark-console-service/src/main/java/org/apache/streampark/console/flow/component/stopsComponent/domain/StopsComponentDomain.java
@@ -1,22 +1,23 @@
 package org.apache.streampark.console.flow.component.stopsComponent.domain;
 
+import org.apache.streampark.console.flow.base.utils.LoggerUtil;
+import org.apache.streampark.console.flow.component.dataSource.mapper.DataSourceMapper;
+import org.apache.streampark.console.flow.component.stopsComponent.entity.StopsComponent;
+import org.apache.streampark.console.flow.component.stopsComponent.entity.StopsComponentGroup;
+import org.apache.streampark.console.flow.component.stopsComponent.entity.StopsComponentProperty;
+import org.apache.streampark.console.flow.component.stopsComponent.mapper.StopsComponentGroupMapper;
+import org.apache.streampark.console.flow.component.stopsComponent.mapper.StopsComponentMapper;
+import org.apache.streampark.console.flow.component.stopsComponent.mapper.StopsComponentPropertyMapper;
+import org.apache.streampark.console.flow.component.stopsComponent.vo.StopsComponentVo;
 import java.util.Arrays;
 import java.util.List;
-import javax.annotation.Resource;
 import org.apache.commons.lang3.StringUtils;
 import org.slf4j.Logger;
+import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Component;
 import org.springframework.transaction.annotation.Isolation;
 import org.springframework.transaction.annotation.Propagation;
 import org.springframework.transaction.annotation.Transactional;
-import org.apache.streampark.console.flow.base.util.LoggerUtil;
-import org.apache.streampark.console.flow.component.stopsComponent.mapper.StopsComponentGroupMapper;
-import org.apache.streampark.console.flow.component.stopsComponent.mapper.StopsComponentMapper;
-import org.apache.streampark.console.flow.component.stopsComponent.mapper.StopsComponentPropertyMapper;
-import org.apache.streampark.console.flow.component.stopsComponent.model.StopsComponent;
-import org.apache.streampark.console.flow.component.stopsComponent.model.StopsComponentGroup;
-import org.apache.streampark.console.flow.component.stopsComponent.model.StopsComponentProperty;
-import org.apache.streampark.console.flow.component.stopsComponent.vo.StopsComponentVo;
 
 @Component
 @Transactional(
@@ -26,13 +27,25 @@ import org.apache.streampark.console.flow.component.stopsComponent.vo.StopsCompo
     rollbackFor = Exception.class)
 public class StopsComponentDomain {
 
-  Logger logger = LoggerUtil.getLogger();
+  /** Introducing logs, note that they are all packaged under "org.slf4j" */
+  private final Logger logger = LoggerUtil.getLogger();
 
-  @Resource private StopsComponentMapper stopsComponentMapper;
+  private final StopsComponentPropertyMapper stopsComponentPropertyMapper;
+  private final StopsComponentGroupMapper stopsComponentGroupMapper;
+  private final StopsComponentMapper stopsComponentMapper;
+  private final DataSourceMapper dataSourceMapper;
 
-  @Resource private StopsComponentPropertyMapper stopsComponentPropertyMapper;
-
-  @Resource private StopsComponentGroupMapper stopsComponentGroupMapper;
+  @Autowired
+  public StopsComponentDomain(
+      StopsComponentPropertyMapper stopsComponentPropertyMapper,
+      StopsComponentGroupMapper stopsComponentGroupMapper,
+      StopsComponentMapper stopsComponentMapper,
+      DataSourceMapper dataSourceMapper) {
+    this.stopsComponentPropertyMapper = stopsComponentPropertyMapper;
+    this.stopsComponentGroupMapper = stopsComponentGroupMapper;
+    this.stopsComponentMapper = stopsComponentMapper;
+    this.dataSourceMapper = dataSourceMapper;
+  }
 
   public int addListStopsComponentAndChildren(List<StopsComponent> stopsComponentList) {
     if (null == stopsComponentList || stopsComponentList.size() <= 0) {
@@ -60,6 +73,15 @@ public class StopsComponentDomain {
           stopsComponentPropertyMapper.insertStopsComponentProperty(properties);
       insertRows += insertStopsTemplateRows;
     }
+    // Change the corresponding "datasource" data source to available
+    if (stopsComponent.getIsDataSource()) {
+      dataSourceMapper.updateDataSourceIsAvailableByBundle(1, stopsComponent.getBundle());
+      // Modify image url of "datasource"
+      dataSourceMapper.updateDataSourceImageUrlByBundle(
+          stopsComponent.getBundle(), stopsComponent.getImageUrl());
+    } else {
+      dataSourceMapper.updateDataSourceIsAvailableByBundle(0, stopsComponent.getBundle());
+    }
     return insertRows;
   }
 
@@ -67,17 +89,19 @@ public class StopsComponentDomain {
     if (null == stopsComponent) {
       return 0;
     }
-    List<StopsComponentProperty> properties = stopsComponent.getProperties();
-    if (null == properties || properties.size() <= 0) {
-      return 0;
-    }
-    int deleteStopsComponentPropertyCount =
+    //        int affectedRows = 0;
+    //        List<StopsComponentProperty> properties = stopsComponent.getProperties();
+    //        if (null != properties && properties.size() > 0) {
+    //            affectedRows =
+    // stopsComponentPropertyMapper.deleteStopsComponentPropertyByStopId(stopsComponent.getId());
+    //        }
+    int affectedRows =
         stopsComponentPropertyMapper.deleteStopsComponentPropertyByStopId(stopsComponent.getId());
 
     int deleteStopsComponentRows =
         stopsComponentMapper.deleteStopsComponentById(stopsComponent.getId());
-
-    return deleteStopsComponentRows + deleteStopsComponentPropertyCount;
+    affectedRows += deleteStopsComponentRows;
+    return affectedRows;
   }
 
   public int addListStopsComponent(List<StopsComponent> stopsComponentList) {
@@ -96,13 +120,6 @@ public class StopsComponentDomain {
       return 0;
     }
     return stopsComponentMapper.insertStopsComponent(stopsComponent);
-  }
-
-  public int deleteStopsComponent() {
-    // empty the "Stop" message and insert
-    stopsComponentPropertyMapper.deleteStopsComponentProperty();
-    int deleteRows = stopsComponentMapper.deleteStopsComponent();
-    return deleteRows;
   }
 
   public StopsComponent getStopsComponentById(String id) {
@@ -125,9 +142,11 @@ public class StopsComponentDomain {
     for (StopsComponentGroup stopGroup : stopsComponentGroupList) {
       String stopGroupId = stopGroup.getId();
       String stopsTemplateId = stopsComponent.getId();
+      String engineType = stopsComponent.getEngineType();
+
       int insertAssociationGroupsStopsTemplate =
           stopsComponentGroupMapper.insertAssociationGroupsStopsTemplate(
-              stopGroupId, stopsTemplateId);
+              stopGroupId, stopsTemplateId, engineType);
       affectedRows += insertAssociationGroupsStopsTemplate;
       logger.info(
           "association_groups_stops_template Association table insertion affects the number of rows : "
@@ -136,19 +155,29 @@ public class StopsComponentDomain {
     return affectedRows;
   }
 
+  public int deleteStopsComponentByEngineType(String engineType) {
+    stopsComponentPropertyMapper.deleteStopsComponentProperty();
+    return stopsComponentMapper.deleteStopsComponent(engineType);
+  }
+
   public int deleteStopsComponent(StopsComponent stopsComponent) {
+    if (null == stopsComponent) {
+      return 0;
+    }
     // delete relationship
     stopsComponentGroupMapper.deleteGroupCorrelationByStopId(stopsComponent.getId());
     logger.debug("Successful delete " + stopsComponent.getName() + " 's association!!!");
 
     // delete stop
-    int stopCount = deleteStopsComponentAndChildren(stopsComponent);
+    int affectedRows = deleteStopsComponentAndChildren(stopsComponent);
     logger.debug("Successful delete " + stopsComponent.getName() + " !!!");
 
     // delete group
     String[] stopsComponentGroup = stopsComponent.getGroups().split(",");
     List<StopsComponentGroup> stopsComponentGroupList =
-        stopsComponentGroupMapper.getStopGroupByNameList(Arrays.asList(stopsComponentGroup));
+        stopsComponentGroupMapper.getStopGroupByNameList(
+            Arrays.asList(stopsComponentGroup), stopsComponent.getEngineType());
+
     for (StopsComponentGroup sGroup : stopsComponentGroupList) {
 
       int count = stopsComponentGroupMapper.getGroupStopCount(sGroup.getId());
@@ -157,15 +186,15 @@ public class StopsComponentDomain {
         logger.debug("Successful delete " + stopsComponent.getName() + " group!!!");
       }
     }
-
-    return stopCount;
+    // Change the corresponding "datasource" data source to unavailable
+    dataSourceMapper.updateDataSourceIsAvailableByBundle(0, stopsComponent.getBundle());
+    return affectedRows;
   }
 
   /**
    * getStopsComponentByBundle
    *
-   * @param bundle
-   * @return
+   * @param bundle bundle
    */
   public StopsComponent getStopsComponentByBundle(String bundle) {
     return stopsComponentMapper.getStopsComponentByBundle(bundle);
@@ -196,8 +225,9 @@ public class StopsComponentDomain {
     return stopsComponentGroupMapper.insertStopGroup(stopsComponentGroup);
   }
 
-  public List<StopsComponentGroup> getStopGroupByNameList(List<String> groupNameList) {
-    return stopsComponentGroupMapper.getStopGroupByNameList(groupNameList);
+  public List<StopsComponentGroup> getStopGroupByNameList(
+      List<String> groupNameList, String engineType) {
+    return stopsComponentGroupMapper.getStopGroupByNameList(groupNameList, engineType);
   }
 
   public StopsComponentGroup getStopsComponentGroupByGroupName(String groupName) {
@@ -212,27 +242,29 @@ public class StopsComponentDomain {
     return stopGroupByName.get(0);
   }
 
-  public int deleteStopsComponentGroup() {
+  public int deleteStopsComponentGroup(String engineType) {
     // the group table information is cleared
     // The call is successful, the group table information is cleared and then
     // inserted.
-    stopsComponentGroupMapper.deleteGroupCorrelation();
-    int deleteRows = stopsComponentGroupMapper.deleteGroup();
+    stopsComponentGroupMapper.deleteGroupCorrelation(engineType);
+    int deleteRows = stopsComponentGroupMapper.deleteGroup(engineType);
     logger.debug("Successful deletion Group" + deleteRows + "piece of data!!!");
     return deleteRows;
   }
 
-  public List<StopsComponentGroup> getStopGroupList() {
-    return stopsComponentGroupMapper.getStopGroupList();
+  public List<StopsComponentGroup> getStopGroupList(String engineType) {
+    return stopsComponentGroupMapper.getStopGroupList(engineType);
   }
 
-  public int insertAssociationGroupsStopsTemplate(String stopGroupId, String stopsTemplateId) {
+  public int insertAssociationGroupsStopsTemplate(
+      String stopGroupId, String stopsTemplateId, String engineType) {
     return stopsComponentGroupMapper.insertAssociationGroupsStopsTemplate(
-        stopGroupId, stopsTemplateId);
+        stopGroupId, stopsTemplateId, engineType);
   }
 
-  public List<StopsComponentGroup> getStopGroupByGroupNameList(List<String> groupName) {
-    return stopsComponentGroupMapper.getStopGroupByGroupNameList(groupName);
+  public List<StopsComponentGroup> getStopGroupByGroupNameList(
+      List<String> groupName, String engineType) {
+    return stopsComponentGroupMapper.getStopGroupByGroupNameList(groupName, engineType);
   }
 
   public int deleteGroupCorrelationByGroupIdAndStopId(String stopGroupId, String stopsTemplateId) {
@@ -242,5 +274,61 @@ public class StopsComponentDomain {
 
   public List<StopsComponentVo> getManageStopsComponentListByGroupId(String stopGroupId) {
     return stopsComponentMapper.getManageStopsComponentListByGroupId(stopGroupId);
+  }
+
+  public List<StopsComponent> getStopsComponentByName(String stopsName) {
+    return stopsComponentMapper.getStopsComponentByName(stopsName);
+  }
+
+  public List<StopsComponent> getDataSourceStopList() {
+    return stopsComponentMapper.getDataSourceStopList();
+  }
+
+  public List<StopsComponentProperty> getDataSourceStopsComponentByBundle(
+      String stopsTemplateBundle) {
+    StopsComponent stopsComponent =
+        stopsComponentMapper.getDataSourceStopsComponentByBundle(stopsTemplateBundle);
+    return stopsComponent.getProperties();
+  }
+
+  public List<StopsComponent> getStopsComponentByStopsHubId(String stopsHubId) {
+    return stopsComponentMapper.getStopsComponentByStopsHubId(stopsHubId);
+  }
+
+  /**
+   * @Description update flow+stops_template @Param stopsComponent @Return int @Author TY @Date
+   * 16:32 2023/4/3
+   */
+  public int updateStopsComponent(StopsComponent stopsComponent) {
+    return stopsComponentMapper.updateStopsComponent(stopsComponent);
+  }
+
+  /**
+   * @Description delete properties @Param stopsId flow_stops_template id @Return int @Author
+   * TY @Date 16:36 2023/4/3
+   */
+  public int deleteStopsComponentProperty(String stopsId) {
+    return stopsComponentPropertyMapper.deleteStopsComponentPropertyByStopId(stopsId);
+  }
+
+  /**
+   * add properties
+   *
+   * @param stopsComponentPropertyList stopsComponentPropertyList
+   */
+  public int insertStopsComponentProperty(List<StopsComponentProperty> stopsComponentPropertyList) {
+    return stopsComponentPropertyMapper.insertStopsComponentProperty(stopsComponentPropertyList);
+  }
+
+  public List<StopsComponent> getSystemDefaultStops(String engineType) {
+    return stopsComponentMapper.getSystemDefaultStops(engineType);
+  }
+
+  public StopsComponent getOnlyStopsComponentByBundle(String bundle) {
+    return stopsComponentMapper.getOnlyStopsComponentByBundle(bundle);
+  }
+
+  public List<StopsComponent> getOnlyStopsComponentByBundles(String[] bundles) {
+    return stopsComponentMapper.getOnlyStopsComponentByBundles(bundles);
   }
 }
